added header and item card to component improved load more logic
