+++ lib/demo_web/store/store.ex
@@ -34,8 +34,8 @@ defmodule DemoWeb.Store do
     end
 
     def getPhonesPerPage(per_page, page) do
-      slice_amount = per_page * page
-      Enum.slice(@phones, 1..slice_amount)
+      slice_amount = per_page * page - 1
+      Enum.slice(@phones, 0..slice_amount)
     end
 
     def phoneCount do
@@ -43,11 +43,14 @@ defmodule DemoWeb.Store do
     end
 
     def getPhones do
-      Enum.slice(@phones, 1..5)
+      Enum.slice(@phones, 0..3)
     end
 
     def getAllPhones do
       @phones
     end
 
+    def showLoadMore(per_page, page) do
+      per_page * page < length(@phones)
+    end
 end
